crypto_ai_bot/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-telegram-bot>=20.0
python-binance>=1.0.19
pandas>=2.0.0
pandas-ta>=0.3.14b
mplfinance>=0.12.10b0
google-generativeai>=0.7.0
numba
matplotlib
python-dotenv
httpx
